+++ script/staking/handler.go
@@ -37,10 +37,10 @@ var (
 	errUpdateForeverBucket  = errors.New("can't update forever bucket")
 
 	// amount
-	errLessThanMinimalBalance  = errors.New("amount less than minimal balance (" + new(big.Int).Div(MIN_REQUIRED_BY_DELEGATE, big.NewInt(1e18)).String() + " MTRG)")
-	errLessThanMinBoundBalance = errors.New("amount less than minimal balance (" + new(big.Int).Div(MIN_BOUND_BALANCE, big.NewInt(1e18)).String() + " MTRG)")
-	errNotEnoughMTR            = errors.New("not enough MTR")
-	errNotEnoughMTRG           = errors.New("not enough MTRG")
+	errLessThanMinimalBalance  = errors.New("amount less than minimal balance (" + new(big.Int).Div(MIN_REQUIRED_BY_DELEGATE, big.NewInt(1e18)).String() + " VERSE)")
+	errLessThanMinBoundBalance = errors.New("amount less than minimal balance (" + new(big.Int).Div(MIN_BOUND_BALANCE, big.NewInt(1e18)).String() + " VERSE)")
+	errNotEnoughMTR            = errors.New("not enough STPT")
+	errNotEnoughMTRG           = errors.New("not enough VERSE")
 
 	// candidate
 	errCandidateNotListed          = errors.New("candidate address is not listed")